Don't use ranges to filter out bad file IDs
`filter` calls front() twice for each element on the range it is
passed,  and `map` calls the mapping function every time
front() is called. So by mapping a path to a call to open()
and then filtering it, we call open() twice on each path.

How to solve this is worth debate. Am I at fault for giving my
mapping side-effects? Or can we tweak the standard library ranges
to avoid this problem?