Move more file processing into ranges using cache
cache() returns a range that makes sure the front() of its input
is only called once per element. This is useful for input ranges
that have side effects (such as printing to stderr or opening
file handles, in our cases) when subsequent ranges in the chain
call front() multiple times (filter does this, for example).